fix : 나무 상세 페이지 렌더 지연 개선
